Extracted InMemoryMetadataIO into a separate class and used that in the IncomingMetricMetadataAnalyzerTest
