+++ blueflood-core/src/test/java/com/rackspacecloud/blueflood/inputs/processors/IncomingMetricMetadataAnalyzerTest.java
@@ -1,13 +1,14 @@
 package com.rackspacecloud.blueflood.inputs.processors;
 
 import com.rackspacecloud.blueflood.cache.MetadataCache;
-import com.rackspacecloud.blueflood.exceptions.CacheException;
-import com.rackspacecloud.blueflood.io.IntegrationTestBase;
+
+import com.rackspacecloud.blueflood.io.MetadataIO;
 import com.rackspacecloud.blueflood.service.IncomingMetricMetadataAnalyzer;
 import com.rackspacecloud.blueflood.types.IMetric;
 import com.rackspacecloud.blueflood.types.Locator;
 import com.rackspacecloud.blueflood.types.Metric;
 import com.rackspacecloud.blueflood.types.MetricMetadata;
+import com.rackspacecloud.blueflood.utils.InMemoryMetadataIO;
 import com.rackspacecloud.blueflood.utils.TimeValue;
 import junit.framework.Assert;
 import org.junit.Test;
@@ -15,17 +16,23 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.concurrent.TimeUnit;
 
-public class IncomingMetricMetadataAnalyzerTest extends IntegrationTestBase{
+public class IncomingMetricMetadataAnalyzerTest {
+    private final MetadataIO metadataIO;
+
+    public IncomingMetricMetadataAnalyzerTest () {
+        this.metadataIO = new InMemoryMetadataIO();
+    }
 
     @Test
     public void test_ScanMetrics_DoesNotStoreTypeAndUnit_ForNumericMetrics() throws Exception {
         MetadataCache cache = MetadataCache.createLoadingCacheInstance(new TimeValue(5, TimeUnit.MINUTES), 1);
+        cache.setIO(this.metadataIO);
         IncomingMetricMetadataAnalyzer analyzer = new IncomingMetricMetadataAnalyzer(cache);
 
         ArrayList<IMetric> metricsList = new ArrayList<IMetric>();
 
         Locator locator = Locator.createLocatorFromDbKey("integer_metric");
-        metricsList.add(getRandomIntMetric(locator, 1));
+        metricsList.add(new Metric(locator, 123, 14200000, new TimeValue(1, TimeUnit.DAYS),"somethings"));
 
         analyzer.scanMetrics(metricsList);
         String type = cache.get(locator, MetricMetadata.TYPE.name().toLowerCase());
@@ -38,18 +45,19 @@ public class IncomingMetricMetadataAnalyzerTest extends IntegrationTestBase{
     @Test
     public void test_ScanMetricS_StoresTypeAndUnit_ForStringMetrics() throws  Exception {
         MetadataCache cache = MetadataCache.createLoadingCacheInstance(new TimeValue(5, TimeUnit.MINUTES), 1);
+        cache.setIO(this.metadataIO);
         IncomingMetricMetadataAnalyzer analyzer = new IncomingMetricMetadataAnalyzer(cache);
 
         ArrayList<IMetric> metricsList = new ArrayList<IMetric>();
 
         Locator locator = Locator.createLocatorFromDbKey("string_metric");
-        metricsList.add(getRandomStringmetric(locator, 1));
+        metricsList.add(new Metric(locator, "verystringy", 142000001, new TimeValue(1, TimeUnit.DAYS),"somethings"));
 
         analyzer.scanMetrics(metricsList);
         String type = cache.get(locator, MetricMetadata.TYPE.name().toLowerCase());
         Assert.assertEquals("S", type);
 
         String unit = cache.get(locator, MetricMetadata.UNIT.name().toLowerCase());
-        Assert.assertEquals("unknown", unit);
+        Assert.assertEquals("somethings", unit);
     }
 }
